+++ app/Http/Controllers/PhoneOperatorController.php
@@ -2,7 +2,9 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Country;
 use App\Models\PhoneOperator;
+use App\Models\StartNumber;
 use Exception;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
@@ -12,21 +14,26 @@ class PhoneOperatorController extends Controller
 {
     public function index()
     {
-        $phoneOperators = PhoneOperator::orderBy('wording')->get();
+        $this->authorize('ROLE_PHONE_OPERATOR_READ', PhoneOperator::class);
+        $phoneOperators = PhoneOperator::with('startNumbers')->orderBy('wording')->get();
+        $countries = Country::orderBy('name_fr')->get();
         return new JsonResponse([
-            'datas' => ['phoneOperators' => $phoneOperators]
+            'datas' => ['phoneOperators' => $phoneOperators, 'countries' => $countries]
         ], 200);
     }
 
     public function store(Request $request)
     {
+        $this->authorize('ROLE_PHONE_OPERATOR_CREATE', PhoneOperator::class);
         $this->validate(
             $request,
             [
+                'country' => 'required',
                 'wording' => 'required|unique:phone_operators|max:150',
                 'description' => 'max:255',
             ],
             [
+                'country.required' => "Le choix d'un pays est obligatoire.",
                 'wording.required' => "Le libellé est obligatoire.",
                 'wording.unique' => "Cet opérateur téléphonique existe déjà.",
                 'wording.max' => "Le libellé ne doit pas dépasser 150 caractères.",
@@ -39,7 +46,20 @@ class PhoneOperatorController extends Controller
             $phoneOperator->code = Str::random(10);
             $phoneOperator->wording = $request->wording;
             $phoneOperator->description = $request->description;
+            $phoneOperator->country_id = $request->country;
             $phoneOperator->save();
+
+            $startNumbers = [];
+            if ($request->startNumbers) {
+                foreach ($request->startNumbers as $key => $number) {
+                    $startNumber = new StartNumber();
+                    $startNumber->number = $number;
+                    $startNumber->phone_operator_id = $phoneOperator->id;
+                    $startNumber->save();
+
+                    array_push($startNumbers, $startNumber);
+                }
+            }
 
             $success = true;
             $message = "Enregistrement effectué avec succès.";
@@ -47,8 +67,10 @@ class PhoneOperatorController extends Controller
                 'phoneOperator' => $phoneOperator,
                 'success' => $success,
                 'message' => $message,
+                'datas' => ['startNumbers' => $startNumbers]
             ], 200);
         } catch (Exception $e) {
+            dd($e);
             $success = false;
             $message = "Erreur survenue lors de l'enregistrement.";
             return new JsonResponse([
@@ -60,24 +82,52 @@ class PhoneOperatorController extends Controller
 
     public function update(Request $request, $id)
     {
+        $this->authorize('ROLE_PHONE_OPERATOR_UPDATE', PhoneOperator::class);
         $phoneOperator = PhoneOperator::findOrFail($id);
         $this->validate(
             $request,
             [
+                'country' => 'required',
                 'wording' => 'required|max:150',
                 'description' => 'max:255',
             ],
             [
+                'country.required' => "Le choix d'un pays est obligatoire.",
                 'wording.required' => "Le libellé est obligatoire.",
                 'wording.max' => "Le libellé ne doit pas dépasser 150 caractères.",
                 'description.max' => "La description ne doit pas dépasser 255 caractères."
             ]
         );
 
+        $existingPhoneOperators = PhoneOperator::where('wording', $request->wording)->get();
+        if (!empty($existingPhoneOperators) && sizeof($existingPhoneOperators) >= 1) {
+            $success = false;
+            return new JsonResponse([
+                'success' => $success,
+                'existingPhoneOperator' => $existingPhoneOperators[0],
+                'message' => "L'opérateur téléphonique " . $existingPhoneOperators[0]->wording . " existe déjà"
+            ], 200);
+        }
+
         try {
             $phoneOperator->wording = $request->wording;
             $phoneOperator->description = $request->description;
+            $phoneOperator->country_id = $request->country;
             $phoneOperator->save();
+
+            StartNumber::where('phone_operator_id', $phoneOperator->id)->delete();
+
+            $startNumbers = [];
+            if ($request->startNumbers) {
+                foreach ($request->startNumbers as $key => $number) {
+                    $startNumber = new StartNumber();
+                    $startNumber->number = $number;
+                    $startNumber->phone_operator_id = $phoneOperator->id;
+                    $startNumber->save();
+
+                    array_push($startNumbers, $startNumber);
+                }
+            }
 
             $success = true;
             $message = "Modification effectuée avec succès.";
@@ -85,6 +135,7 @@ class PhoneOperatorController extends Controller
                 'phoneOperator' => $phoneOperator,
                 'success' => $success,
                 'message' => $message,
+                'datas' => ['startNumbers' => $startNumbers]
             ], 200);
         } catch (Exception $e) {
             $success = false;
@@ -98,6 +149,7 @@ class PhoneOperatorController extends Controller
 
     public function destroy($id)
     {
+        $this->authorize('ROLE_PHONE_OPERATOR_DELETE', PhoneOperator::class);
         $phoneOperator = PhoneOperator::findOrFail($id);
         try {
             $phoneOperator->delete();
@@ -116,5 +168,14 @@ class PhoneOperatorController extends Controller
                 'message' => $message,
             ], 400);
         }
+    }
+
+    public function show($id)
+    {
+        $this->authorize('ROLE_PHONE_OPERATOR_READ', PhoneOperator::class);
+        $phoneOperator = PhoneOperator::with('startNumbers')->findOrFail($id);
+        return new JsonResponse([
+            'phoneOperator' => $phoneOperator
+        ], 200);
     }
 }
