+++ app/Http/Controllers/UserController.php
@@ -13,6 +13,7 @@ class UserController extends Controller
 {
     public function index()
     {
+        $this->authorize('ROLE_USER_READ', User::class);
         $users = User::orderBy('last_name')->orderBy('first_name')->get();
         $userTypes = UserType::orderBy('wording')->get();
 
@@ -23,27 +24,32 @@ class UserController extends Controller
 
     public function store(Request $request)
     {
-        $this->validate($request, [
-            'email' => 'required|email',
-            'password' => 'required|min:8',
-            'last_name' => 'required|max:30',
-            'first_name' => 'required|max:50',
-            'date_of_birth' => 'required|date|date_format:Ymd|before:today',
-            'place_of_birth' => 'required|max:100',
-        ], [
-            'email.required' => "Le champ email est obligatoire.",
-            'email.email' => "Le champ email est incorrect.",
-            'password.required' => "Le champ mot de passe est obligatoire.",
-            'password.min' => "Le champ mot de passe doit contenir 8 caractères.",
-            'last_name.required' => "e champ nom est obligatoire.",
-            'last_name.max' => "Le champ nom ne doit pas dépasser 30 caractères.",
-            'first_name.required' => "Le champ prénom(s) est obligatoire.",
-            'first_name.max' => "Le champ prénom(s) ne doit pas dépasser 50 caractères.",
-            'date_of_birth.required' => "Le champ date de naissance est obligatoire.",
-            'date_of_birth.date' => "Le champ date de naissance doit être une date valide.",
-            'date_of_birth.date_format' => "Le champ date de naissance doit être sous le format : Année Mois Jour.",
-            'date_of_birth.before' => "Le champ date de naissance doit être antérieure ou égale à aujourd'hui.",
-        ]);
+        $this->authorize('ROLE_USER_CREATE', User::class);
+        $this->validate(
+            $request,
+            [
+                'email' => 'required|email',
+                'password' => 'required|min:8',
+                'last_name' => 'required|max:30',
+                'first_name' => 'required|max:50',
+                'date_of_birth' => 'required|date|date_format:Ymd|before:today',
+                'place_of_birth' => 'required|max:100',
+            ],
+            [
+                'email.required' => "Le champ email est obligatoire.",
+                'email.email' => "Le champ email est incorrect.",
+                'password.required' => "Le champ mot de passe est obligatoire.",
+                'password.min' => "Le champ mot de passe doit contenir 8 caractères.",
+                'last_name.required' => "e champ nom est obligatoire.",
+                'last_name.max' => "Le champ nom ne doit pas dépasser 30 caractères.",
+                'first_name.required' => "Le champ prénom(s) est obligatoire.",
+                'first_name.max' => "Le champ prénom(s) ne doit pas dépasser 50 caractères.",
+                'date_of_birth.required' => "Le champ date de naissance est obligatoire.",
+                'date_of_birth.date' => "Le champ date de naissance doit être une date valide.",
+                'date_of_birth.date_format' => "Le champ date de naissance doit être sous le format : Année Mois Jour.",
+                'date_of_birth.before' => "Le champ date de naissance doit être antérieure ou égale à aujourd'hui.",
+            ]
+        );
 
         $existingUser = UserType::where('last_name', $request->last_name)->where('first_name', $request->first_name)->get();
         if (!empty($existingUser) && sizeof($existingUser) >= 1) {
@@ -86,28 +92,33 @@ class UserController extends Controller
 
     public function update(Request $request, $id)
     {
+        $this->authorize('ROLE_USER_UPDATE', User::class);
         $user = User::findOrFail($id);
-        $this->validate($request, [
-            'email' => 'required|email',
-            'password' => 'required|min:8',
-            'last_name' => 'required|max:30',
-            'first_name' => 'required|max:50',
-            'date_of_birth' => 'required|date|date_format:Ymd|before:today',
-            'place_of_birth' => 'required|max:100',
-        ], [
-            'email.required' => "Le champ email est obligatoire.",
-            'email.email' => "Le champ email est incorrect.",
-            'password.required' => "Le champ mot de passe est obligatoire.",
-            'password.min' => "Le champ mot de passe doit contenir 8 caractères.",
-            'last_name.required' => "e champ nom est obligatoire.",
-            'last_name.max' => "Le champ nom ne doit pas dépasser 30 caractères.",
-            'first_name.required' => "Le champ prénom(s) est obligatoire.",
-            'first_name.max' => "Le champ prénom(s) ne doit pas dépasser 50 caractères.",
-            'date_of_birth.required' => "Le champ date de naissance est obligatoire.",
-            'date_of_birth.date' => "Le champ date de naissance doit être une date valide.",
-            'date_of_birth.date_format' => "Le champ date de naissance doit être sous le format : Année Mois Jour.",
-            'date_of_birth.before' => "Le champ date de naissance doit être antérieure ou égale à aujourd'hui.",
-        ]);
+        $this->validate(
+            $request,
+            [
+                'email' => 'required|email',
+                'password' => 'required|min:8',
+                'last_name' => 'required|max:30',
+                'first_name' => 'required|max:50',
+                'date_of_birth' => 'required|date|date_format:Ymd|before:today',
+                'place_of_birth' => 'required|max:100',
+            ],
+            [
+                'email.required' => "Le champ email est obligatoire.",
+                'email.email' => "Le champ email est incorrect.",
+                'password.required' => "Le champ mot de passe est obligatoire.",
+                'password.min' => "Le champ mot de passe doit contenir 8 caractères.",
+                'last_name.required' => "e champ nom est obligatoire.",
+                'last_name.max' => "Le champ nom ne doit pas dépasser 30 caractères.",
+                'first_name.required' => "Le champ prénom(s) est obligatoire.",
+                'first_name.max' => "Le champ prénom(s) ne doit pas dépasser 50 caractères.",
+                'date_of_birth.required' => "Le champ date de naissance est obligatoire.",
+                'date_of_birth.date' => "Le champ date de naissance doit être une date valide.",
+                'date_of_birth.date_format' => "Le champ date de naissance doit être sous le format : Année Mois Jour.",
+                'date_of_birth.before' => "Le champ date de naissance doit être antérieure ou égale à aujourd'hui.",
+            ]
+        );
 
         $existingUser = UserType::where('last_name', $request->last_name)->where('first_name', $request->first_name)->get();
         if (!empty($existingUser) && sizeof($existingUser) >= 1) {
@@ -148,6 +159,7 @@ class UserController extends Controller
 
     public function destroy($id)
     {
+        $this->authorize('ROLE_USER_DELETE', User::class);
         $user = User::findOrFail($id);
         try {
             $user->delete();
@@ -170,6 +182,7 @@ class UserController extends Controller
 
     public function userTypeConfiguration(Request $request, $id)
     {
+        $this->authorize('ROLE_USER_READ', User::class);
         $user = User::findOrFail($id);
         $userType = UserType::findOrFail($request->user_type);
 
@@ -197,6 +210,7 @@ class UserController extends Controller
 
     public function rolesConfiguration(Request $request, $id)
     {
+        $this->authorize('ROLE_USER_READ', User::class);
         $user = User::findOrFail($id);
 
         try {
